requirements.txt
google-adk
pydantic
typing-extensions
pytest
pytest-asyncio
ruff
coverage
